Create pooltool.ani.modes module, where each mode gets its own .py file. Importing pooltool.ani.modes gives access to each of the game modes. Each mode is a child class of `Mode`, which is defined in the __init__.py. Integrity of all classes that inherit `Mode` is carried out in the __init__.py by using ABC module and making dummy instances in the __init__
